s3839908-a2/Controllers/HomeController.cs
using Data;
using Microsoft.AspNetCore.Mvc;
using s3839908_a2.Models;
using SimpleHashing.Net;
using System.Diagnostics;

namespace McbaExample.Controllers;

// Bonus Material: Implement global authorisation check.
//[AllowAnonymous]
public class HomeController : Controller
{
    private static readonly ISimpleHash s_simpleHash = new SimpleHash();

    private readonly McbaContext _context;
    public IActionResult Index() => View();

    public IActionResult Privacy() => View();

    [ResponseCache(Duration = 0, Location = ResponseCacheLocation.None, NoStore = true)]
    public IActionResult Error() =>
        View(new Models.ErrorViewModel { RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier });

    public HomeController(McbaContext context)
    {
        _context = context;
    }

    [HttpPost]
    public async Task<IActionResult> Login(string loginID, string password)
    {
        var login = await _context.Logins.FindAsync(loginID);

        if (login == null || string.IsNullOrEmpty(password) || !s_simpleHash.Verify(password, login.PasswordHash))
        {
            ModelState.AddModelError("LoginFailed", "Login failed, please try again.");
            return View("Index", new Login { LoginID = loginID });
        }

        if (login.Locked)
        {
            ModelState.AddModelError("LoginFailed", "Your account is locked.");
            return View("Index", new Login { LoginID = loginID });
        }

        // Login customer.
        HttpContext.Session.SetInt32(nameof(Customer.CustomerID), login.CustomerID);
        HttpContext.Session.SetString(nameof(Customer.Name), login.Customer.Name);

        return RedirectToAction("Index", "Customer");
    }

    [Route("LogoutNow")]
    public IActionResult Logout()
    {
        // Logout customer.
        HttpContext.Session.Clear();

        return RedirectToAction("Index", "Home");
    }
}
